fix: resolved circular dependencies
