+++ chain-api/src/utils/dex/dexTypes.ts
@@ -14,13 +14,6 @@
  */
 import BigNumber from "bignumber.js";
 
-export type TokenInstanceKeyData = {
-  collection: string;
-  category: string;
-  type: string;
-  additionalKey: string;
-};
-
 export type SwapState = {
   // the amount remaining to be swapped in/out of the input/output asset
   amountSpecifiedRemaining: BigNumber;
@@ -73,6 +66,6 @@ type UserPositionProps = {
   liquidity: string;
 };
 
-type PositionInPool = {
+export type PositionInPool = {
   [key: string]: UserPositionProps[];
 };
