+++ chaincode/src/utils/index.ts
@@ -18,7 +18,6 @@ export * from "./state";
 export * from "./utils";
 export * from "./authorityOrgName";
 export * from "./dexUtils";
-export * from "./finaliseSale";
 export * from "./launchpadSaleUtils";
 
 export { GalaJSONSerializer };
